requirements.txt
pytorch==2.4.1
transformers==4.45.2
peft==0.12.0
fastapi==0.112.2
uvicorn==0.32.1
sentencepiece==0.2.0
protobuf==5.29.2
adapters==1.0.1
